Fix to not to show Datepicker UI when input element is readonly
In scenarios where the input field is not disabled but made readonly,
so the form field can be submitted to the server, Datepicker shows the
Picker UI and allows user to edit the date, that should not be
allowed.

The form input filed could be updated to be editable based on business
logic, then Picker UI should displayed on click event to allow user to
edit the input.

This PR checks if the input field has readonly attribute, then prevent
Picker UI to not show; as well adds a unit test case to check this scenario.